+++ src/main/java/com/alvaro/empresas/passagens/dtos/precios/PrecioDTOResponseViaje.java
@@ -3,12 +3,13 @@ package com.alvaro.empresas.passagens.dtos.precios;
 import com.alvaro.empresas.passagens.autobuses.dtos.pisos.PisoDTOResponse;
 import com.alvaro.empresas.passagens.models.PrecioModel;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public record PrecioDTOResponseViaje(
         UUID id,
-        Float precio,
+        BigDecimal precio,
         Integer nPiso,
         Boolean lleno,
         Integer nSillasDisponibles,
