+++ src/main/java/com/alvaro/empresas/passagens/dtos/precios/PrecioDTOUpdate.java
@@ -2,8 +2,10 @@ package com.alvaro.empresas.passagens.dtos.precios;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record PrecioDTOUpdate(
         @NotNull
-        Float precio
+        BigDecimal precio
 ) {
 }
